Add a tool for getting test paths from ids
Sometimes one has a list of test ids (e.g. from a test run) and wants
to know which files correspond to those tests (e.g. because one has
data stored on a per-file basis). This adds a small command line tool
to get the files corresponding to a test out of the manifest.