[sms] Implement AbortController for SMSReceiver API
This change allows developers to use the SMS Receiver API with the Abort
Controller to cancel once the API is called. This will help developers with
flows for retrying and sending a new code and they want to abort the previous
call they made to the API. If the InfoBar is up during abort, it will
leave the InfoBar open while returning an AbortError to the website. If
another request is made while the InfoBar for 1st request is still open (after
abort), upon clicking 'Verify' the SMS retrieved for the 1st request will
return to the 2nd request.

Video(googlers only):
https://drive.google.com/file/d/1FDXv-Zhconiy5xEg-AJjZDCL3z7P5Z_t/view?usp=sharing

Bug: 976401
Change-Id: I132800f2954fdc370172dcd03789d3f08bcc4040