Add getTransform function to offscreen canvases
It seems like maybe this was just overlooked? The fix is as easy as
adding it to the idl file.

It seems like this is already in the spec as:
https://html.spec.whatwg.org/multipage/canvas.html#offscreencanvasrenderingcontext2d
contains the line
OffscreenCanvasRenderingContext2D includes CanvasTransform;
And https://html.spec.whatwg.org/multipage/canvas.html#canvastransform lists the
getTransform() function.

Bug: 1016454
Change-Id: Icb5405d716524e2a0b6935db81b18ff53cd31d4f