[Trusted Types] Allow duplicated policy name for "*" CSP directive.
This adapts TT to the latest spec, which asks for duplicate policy names to
be accepted when "trusted-types *" is given.

Since this requires different error conditions for duplicate names depending
on the active policy, this CL pushes the checking into the CSP class, rather
than handling it at the calling side.

Bug: 991129
Change-Id: Ia40f84093200a1a6e0fae7b65d080c1b5ab8c2fe